requirements.txt
django==5.0.7
djangorestframework==3.15.2
pillow==10.4.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
psycopg[binary]
